fix: add missing includes and fix type conversion warnings
- Add missing <algorithm> include in CharacterManager.cpp
- Cast strlen result to short in CSProtocol.cpp to fix type conversion warning
- Add explicit cast for WCoinC delta in CashShop.cpp to prevent sign mismatch
- Store c_str() result in variable before use in JoinServerProtocol.cpp
- Update solution files to Visual Studio 2022 format
- Update project files to use v143 toolset and add security warning disables
- Change output directories to local Release folder

+++ DataServer/DataServer/CSProtocol.cpp
@@ -542,7 +542,7 @@ void FriendMemoReadReq(FHP_FRIEND_MEMO_RECV_REQ* lpMsg, int index)
 
 	gQueryManager.GetAsBinary("Memo", (BYTE*)pMsg.Memo, sizeof(pMsg.Memo));
 
-	pMsg.MemoSize = strlen(pMsg.Memo);
+	pMsg.MemoSize = (short)strlen(pMsg.Memo);
 
 	gQueryManager.Close();
